Cambios Servicios Articulo y Articulo Inventarios

+++ bambu/src/main/java/com/panda/bambu/service/inventory/BalanceService.java
@@ -27,18 +27,29 @@ public class BalanceService {
     public List<Balance> findAll(){
         return balanceRepository.findAll();
     }
+    
+    public boolean isBalanceAlreadyPresent(Balance balance) {
+		Balance balanceFound = balanceRepository.findById(balance.getId()).get();
+		if (balanceFound==null) 
+			return false;
+		return true;
+    }
 
-    public void create(int quantity, double totalCost, Article article){
-          
-        if(quantity >= 0 && totalCost >= 0 && article != null){
-             Balance newBalance = new Balance(quantity,totalCost);
-             newBalance.setArticle(article);
-             balanceRepository.save(newBalance);
+    public boolean create(int quantity, double totalCost, Article article){
+      
+        if(article != null){
+            if(quantity >= 0 && totalCost >= 0 && articleService.isArticleAlreadyPresent(article)){
+                Balance newBalance = new Balance(quantity,totalCost);
+                newBalance.setArticle(article);
+                balanceRepository.save(newBalance);
+                return true;
+            }
         }
+        return false;
 
     }
 
-    public void create(int quantity, double totalCost, Long articleId){
+    public boolean create(int quantity, double totalCost, Long articleId){
         
         Article article = articleService.findById(articleId);
 
@@ -46,10 +57,13 @@ public class BalanceService {
            Balance newBalance = new Balance(quantity,totalCost);
            newBalance.setArticle(article);
            balanceRepository.save(newBalance);
+           return true;
         }
+
+        return false;
     }
 
-    public void create(int quantity, double totalCost, String articleCode){
+    public boolean create(int quantity, double totalCost, String articleCode){
         
         Article article = articleService.findByCode(articleCode);
         
@@ -57,18 +71,33 @@ public class BalanceService {
            Balance newBalance = new Balance(quantity,totalCost);
            newBalance.setArticle(article);
            balanceRepository.save(newBalance);
+           return true;
         }
+
+        return false;
     }
 
-    public void delete(Long id){
-         balanceRepository.deleteById(id);  
+    public boolean delete(Long id){
+
+         if (balanceRepository.existsById(id)){
+             balanceRepository.deleteById(id); 
+             return true;
+         }
+         return false;
+          
     }
 
-    public void delete(Balance balance){
-         balanceRepository.delete(balance);
+    public boolean delete(Balance balance){
+
+        if(balance != null && balanceRepository.findById(balance.getId()) != null){
+            balanceRepository.delete(balance);
+            return true;
+        }
+
+        return false;  
     }
 
-    public void deleteAll(){
+    public void deleteAllBalances(){
          balanceRepository.deleteAll();
     }
